Export EditableTable2 stories from table stories.tsx
