+++ src/components/table/stories.tsx
@@ -1,23 +1,14 @@
-import * as R from 'ramda';
-import { useCallback, useState } from 'react';
+import { useState } from 'react';
 import { StoryObj, Meta } from '@storybook/react';
 
 import { Box } from 'rebass';
 import Table from './index';
-import ReadOnlyTable from './read-only';
 
-import dummyValues from './read-only/dummydata.json';
 import dummyValuesEditable from './editable/dummydata.json';
 import EditableTable from './editable';
 import { Button, Checkbox, Select } from '../../index';
 import BlurInput from './editable/blur-input';
-import {
-  TableCellRenderProps,
-  TableCellType,
-  TableHeader,
-  TableRowComponent,
-} from './type';
-import { EditableTable2 } from './editable/EditableTable2';
+import { TableCellType, TableHeader, TableRowComponent } from './type';
 
 const meta: Meta<typeof Table> = {
   title: 'DataDisplay/Tables/Table',
@@ -55,7 +46,7 @@ const Score = ({ value, isHovered }: any) => (
   </Box>
 );
 
-const headers = [
+export const headers = [
   {
     identifier: { name: 'away_team_id' },
     cellRender: ({ value }: { value: string }) => `${value} ++`,
@@ -75,32 +66,7 @@ const headers = [
   },
 ];
 
-export const ReadOnly: StoryObj<typeof ReadOnlyTable> = {
-  args: {
-    values: dummyValues,
-    columnHeaders: headers,
-  },
-  render: (props) => {
-    return (
-      <Box>
-        <ReadOnlyTable
-          {...props}
-          initialStaticColumn="score"
-          actions={[
-            {
-              label: 'go to stats',
-              handler: (column) => {
-                console.log(`go to stats of ${column}`);
-              },
-            },
-          ]}
-        />
-      </Box>
-    );
-  },
-};
-
-const rowComponents: TableRowComponent[] = [
+export const rowComponents: TableRowComponent[] = [
   {
     identifier: { name: 'away_team_id' },
     render: ({ value, onChange, onBlur }) => (
@@ -223,120 +189,6 @@ export const Editable: StoryObj<typeof EditableTable> = {
           ]}
         />
       </>
-    );
-  },
-};
-
-// converting old data structure to new one
-type DataShape = {
-  identifierName: string;
-  value: any;
-};
-
-const convertOldDataToNew = R.map(
-  R.reduce(
-    (acc, item: DataShape) => ({ ...acc, [item.identifierName]: [item.value] }),
-    {},
-  ),
-);
-const dummyValues2 = convertOldDataToNew(dummyValuesEditable as DataShape[][]);
-const editableTable2Columns = [
-  {
-    accessorKey: 'away_team_id',
-    // cell: since type is input we don't need to explicitly define cell
-  },
-  {
-    accessorKey: 'score',
-    cell: ({ value, onChange, onBlur }: TableCellRenderProps) => {
-      const handleChange = () => {
-        onChange(!(value as boolean));
-      };
-
-      return (
-        <Checkbox
-          ml="8px"
-          checked={value as boolean}
-          onChange={handleChange}
-          variant="gray"
-          onBlur={onBlur}
-        />
-      );
-    },
-  },
-  {
-    accessorKey: 'dummycolumn_test1',
-    cell: ({ value, onChange, onBlur }: TableCellRenderProps) => (
-      <Select
-        value={value as string[]}
-        onChange={onChange}
-        options={['1', '2']}
-        placeholder=""
-        onBlur={onBlur}
-      />
-    ),
-  },
-  {
-    accessorKey: 'dummycolumn_test2',
-    // cell: since type is input we don't need to explicitly define cell
-  },
-  {
-    accessorKey: 'home_team_id',
-    cell: ({ value, onChange }: TableCellRenderProps) => (
-      <Select
-        value={value as string[]}
-        onChange={onChange}
-        options={['1', '2']}
-        placeholder=""
-      />
-    ),
-  },
-];
-
-export const Editable2: StoryObj<typeof EditableTable2> = {
-  args: {
-    hasFreezeButton: true,
-    columns: editableTable2Columns as any,
-    actions: [
-      {
-        label: 'go to stats',
-        handler: (column) => {
-          console.log(`go to stats of ${column}`);
-        },
-      },
-    ],
-  },
-  render: (props: any) => {
-    const [data, setData] = useState(dummyValues2);
-
-    const updateData = useCallback(
-      (rowIndex: number, columnId: string, value: any) => {
-        console.log({ rowIndex, columnId, value });
-        // Skip page index reset until after next rerender
-        setData((old) =>
-          old.map((row, index) => {
-            if (index === rowIndex) {
-              return {
-                ...old[rowIndex]!,
-                [columnId]: value,
-              };
-            }
-            return row;
-          }),
-        );
-      },
-      [],
-    );
-
-    const handleDeleteRow = (rowIdx: number) =>
-      setData((data) => data.filter((x, idx) => idx !== rowIdx));
-
-    return (
-      <EditableTable2
-        data={data}
-        onDeleteRow={handleDeleteRow}
-        updateData={updateData}
-        {...props}
-      />
     );
   },
 };
